backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
imagehash
anthropic
python-multipart
python-dotenv
aiofiles
